Bug Fix for Filtering by Date
The method for filtering files by date has been changed. Instead of parsing the file name from the beginning and splitting it at the underscore ("_"), which failed when there were more than two underscores before the date, the file name is now parsed from the end, where it is guaranteed to contain only the date.